backend/requirements.txt
flask
flask_cors
Pillow
numpy
